+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+- Accept AnalysisOptions as an argument to `Linter#lint` and
+  `Linter#lintPackage`.
 
 ## [3.0.0-pre.4] - 2018-03-27